made /projects/$projectId it's own route
